Rename ethertype.rs to ethtype.rs
This is more consistent.
